feat: transactional logic moved from controller to service
